fixed builds, added friend integration
